Removed 'returns' from input process method
Originally, I put together the summary for the "Process_Input_File" method when it was first created and was supposed to return an integer. Since then, I've changed the method to be a 'void' return type. I forgot to remove the comment that said it returned a type of status code.