"\t" char consistency & comment clarification.
Program.cs and DelimiterProcessor.cs were looking for "\t" and "    " (a tab press) respectively. Set both to "\t" for consistency and universality. Also, added a clarification to the message when a user inputs an invalid file type (i.e. instead of saying the file is "empty", it now says the file is "empty or invalid").